% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSOFT.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{CIBERSORT}
\usage{
CIBERSORT(
  sig_matrix,
  mixture_obj,
  perm = 0,
  QN = TRUE,
  absolute = FALSE,
  abs_method = "sig.score"
)
}
\arguments{
\item{sig_matrix}{matrix, signature matrix, LM22}

\item{mixture_obj}{matrix, mixture matrix, row is gene, column is sample}

\item{perm}{number, default 1000(recommend 1000)}

\item{QN}{logical, default F, RNAseq data: recommend QN = F, otherwise T}

\item{absolute}{absolute logical, if T, return absolute score}

\item{abs_method}{character, default 'sig.score', can be 'no.sumto1'}
}
\value{
a matrix
}
\description{
run cibersort
}
\examples{
\dontrun{
 CIBERSORT(sig_matrix_file = LM22, 
           mixture_obj = exprs, 
           perm = 1000, QN = F, absolute = T, abs_method='sig.score')
}
}
