% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool2.R
\name{combn_sample}
\alias{combn_sample}
\title{Combine method}
\usage{
combn_sample(col.1, col.2, sample1 = 2, sample2 = 2, eset)
}
\arguments{
\item{col.1}{character(n) or numerber(n)}

\item{col.2}{character(n) or numerber(n)}

\item{sample1}{number the sample number of group1}

\item{sample2}{number the sample number of group1}

\item{eset}{data.frame the expr dataframe}
}
\value{
dataframe or matrix
}
\description{
combine coloumn from dataframe or colnames
}
\author{
Jiang
}
