% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSOFT.R
\name{CoreAlg}
\alias{CoreAlg}
\title{Core algorithm}
\usage{
CoreAlg(X, y, absolute, abs_method)
}
\arguments{
\item{X}{}

\item{y}{}

\item{absolute}{}

\item{abs_method}{}
}
\value{
a list
}
\description{
Core algorithm
}
