% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{gencode.v22.annot}
\alias{keggpathway.gmt}
\title{Datasets
gencode.v22.annot contains gene annot info.}
\description{
Datasets
gencode.v22.annot contains gene annot info.

Datasets
keggpathway.gmt contains kegg pathway information
}
\keyword{datasets}
