% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEG_DESeq2.ana}
\alias{DEG_DESeq2.ana}
\title{RNAseq differential analysis DESeq2 003<dds.filter_result>}
\usage{
DEG_DESeq2.ana(dds, pval = 0.05, fdr = 0.1, logfc = 1)
}
\arguments{
\item{dds}{the result of DEG_DESeq2.dds}

\item{pval}{PValue}

\item{fdr}{FDR}

\item{logfc}{log2(fold change)}
}
\value{
list(list(groupdata = dds@colData, resdf=res_df, deg=etSig))
}
\description{
The function is designed to perform differential analysis by DESeq2<dds.filter_result>.
}
\examples{
#
}
