% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEG_DESeq2.pca}
\alias{DEG_DESeq2.pca}
\title{RNAseq differential analysis DESeq2 002<pca>}
\usage{
DEG_DESeq2.pca(dds, outdir)
}
\arguments{
\item{dds}{the result of DEG_DESeq2.dds}

\item{outdir}{output dir, pca plot output dir, if the outdir exit the function will remove it}
}
\value{
no return, pca plot to disk
}
\description{
The function is designed to perform differential analysis by DESeq2<pca>.
}
\examples{
#
}
