% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.R
\name{DEG_GO}
\alias{DEG_GO}
\title{GO analysis}
\usage{
DEG_GO(
  genelist,
  orgdb = "org.Hs.eg.db",
  sigNodes = 20,
  resultdir,
  filemark,
  rapid = FALSE
)
}
\arguments{
\item{genelist}{dataframe, the dataframe must contain "ENTREZID" coloum}

\item{orgdb}{character, org type}

\item{sigNodes}{number, the tree plot nodes}

\item{resultdir}{character, the output dir}

\item{filemark}{character, special part of output filename}
}
\value{
# list, GO analsis object go_list
}
\description{
GO analysis by ENTREZID, plot go tree picture and write to disk, retern the result
}
\examples{
#
}
\author{
Jiang
}
