% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.R
\name{DEG_KEGG}
\alias{DEG_KEGG}
\title{KEGG analysis}
\usage{
DEG_KEGG(genelist, orgdb = "org.Hs.eg.db", org_kegg = "hsa")
}
\arguments{
\item{genelist}{dataframe, the dataframe must contain "UNIPROT" coloum}

\item{orgdb}{character, org type}

\item{org_kegg}{character, kegg org type}
}
\value{
# list, KEGG analsis object & its df, list(KEGG=KEGG, pSigDF=keggdf)
}
\description{
KEGG analysis by UNIPROT, retern the result
}
\examples{
#
}
\author{
Jiang
}
