% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEG_prepareData.DEres}
\alias{DEG_prepareData.DEres}
\title{S3 DEG_prepareData.DEres: prepare the data that LZ diff analysis need}
\usage{
\method{DEG_prepareData}{DEres}(obj, f_mark = "DE")
}
\arguments{
\item{obj}{object}

\item{f_mark}{character a analysis mark(optional), default: "DE"}
}
\value{
list
}
\description{
S3 method of object DEres, prepare the data that LZ diff analysis need, \cr
return a list(eset=xx, group=xx, f_mark="xx"), S3 general function DEG_prepareData.DEres
}
