% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEG_prepareData}
\alias{DEG_prepareData}
\title{S3 DEG_prepareData: general function DEG_prepareData}
\usage{
DEG_prepareData(obj)
}
\arguments{
\item{obj}{object}
}
\value{
list
}
\description{
S3 method initialize, initialize the S3 general function DEG_prepareData()
}
