% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.R
\name{DEG_prepareENRICH}
\alias{DEG_prepareENRICH}
\title{Enrich analysis data prepare}
\usage{
DEG_prepareENRICH(resdf, fc.list, p = 0.05, fdr = 0.1)
}
\arguments{
\item{resdf}{dataframe, the DEG anlysis object(all_father$DIFF.ALL')}

\item{fc.list}{list, name list such as list('1.5' = log2(1.5), '2' = log2(2))}

\item{p}{number, pvale}

\item{fdr}{number, fdr}
}
\value{
list
}
\description{
Transform the DEG anlysis object(all_father$DIFF.ALL') to GO analsis
}
\author{
Jiang
}
