% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.R
\name{DEG_prepareGOglist}
\alias{DEG_prepareGOglist}
\title{GO analysis data prepare}
\usage{
DEG_prepareGOglist(resdf, logfc, p = 0.05, fdr = 0.1)
}
\arguments{
\item{resdf}{dataframe, the DEG anlysis object(all_father$DIFF.ALL')}

\item{logfc}{number, logfc}

\item{p}{number, pvale}

\item{fdr}{number, fdr}
}
\value{
# list(all=all, up=up, down=down)
}
\description{
Transform the DEG anlysis object(all_father$DIFF.ALL') to GO analsis
Note: The go enrich result of the length of different deg genes is different!
}
\examples{
#
}
\author{
Jiang
}
