% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.R
\name{DEG_runGO}
\alias{DEG_runGO}
\title{run GO analysis in batches(upgene, downgene, allgene)}
\usage{
DEG_runGO(outdir, genelist)
}
\arguments{
\item{outdir}{character, the output dir}

\item{genelist}{list, go genelist, result of DEG_prepareGOglist}
}
\value{
# GO object & write result to xlsx
}
\description{
run GO analysis in batches(upgene, downgene, allgene) and write result to xlsx
}
\examples{
#
}
