% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{DEG_runGSEA}
\alias{DEG_runGSEA}
\title{Run GSEA of gmt}
\usage{
DEG_runGSEA(
  genelist,
  gmt_set,
  pic.save,
  outdir,
  filename,
  pic.w = 7,
  pic.h = 6,
  num = 1,
  pvalue = 1
)
}
\arguments{
\item{genelist}{list, a name list store the value of all gene after arrage}

\item{gmt_set}{dataframe or a object, a gmt dataframe filename or gmt object, col1: term, col2: gene
#            term  gene
# GOBP_PYROPTOSIS  AIM2
# GOBP_PYROPTOSIS  APIP
# GOBP_PYROPTOSIS CASP1
# GOBP_PYROPTOSIS CASP4
# ...               ...
# KEGG...          ...}

\item{pic.save}{logical if save the picture, default is F, it's not reccommed set as T if gmt is lots of pathway}

\item{outdir}{character the picture outdir}

\item{filename}{character the picture filename}

\item{pic.w}{number picture width}

\item{pic.h}{number picture height}

\item{num}{number the row postion of result of  clusterProfiler::GSEA()}

\item{pvalue}{number the p value of gsea analysis}
}
\value{
gsea
}
\description{
Run GSEA analysis with a gmt file consist of lots of pathway or a single pathway
}
\author{
Jiang
}
