% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.R
\name{DEG_runKEGG}
\alias{DEG_runKEGG}
\title{run KEGG analysis in batches(upgene, downgene, allgene)}
\usage{
DEG_runKEGG(outdir, genelist)
}
\arguments{
\item{outdir}{character, the output dir}

\item{genelist}{list, go genelist, result of DEG_prepareGOglist}
}
\value{
# KEGG object & write result to xlsx
}
\description{
run KEGG analysis in batches(upgene, downgene, allgene) and write result to xlsx
}
\examples{
#
}
