% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEG_tw.test}
\alias{DEG_tw.test}
\title{Title}
\usage{
DEG_tw.test(
  eset,
  group,
  method = "wilcox.test",
  pval = 0.05,
  fdr = 0.1,
  logfc = log2(2)
)
}
\arguments{
\item{eset}{data.frame, the exprs data, the data should not be log2 transformed}

\item{group}{data.frame, the pheno information, the first coloumn is the type information and should be factor.}

\item{method}{character, the method of test, default is wilcox.test, alternative is t.test}

\item{pval}{number, the pvalue cutoff}

\item{fdr}{number, the fdr cutoff}

\item{logfc}{number, the log2(fc) value cutoff}
}
\value{
list(resdf = res, deg = deg)
}
\description{
Title
}
\examples{
\dontrun{
pn_eset <- data.table::fread("./protein_count_pn.csv", data.table = F)
pn_eset <- pn_eset \%>\% tibble::column_to_rownames(var = "Protein")

group <- read.csv("./group.csv", row.names = 1)
group$Type <- as.factor(group$Type)

diff <- DEG_tw.test(eset = df, group = group)
}
}
\author{
Jiang
}
