% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.plot.R
\name{DEG_venn}
\alias{DEG_venn}
\title{Venn Plot}
\usage{
DEG_venn(datalist, transparency = 0.6)
}
\arguments{
\item{datalist}{list, list(diffan$deg$symbol, diffagek)}

\item{transparency}{number, transparency}
}
\value{
#
}
\description{
Plot Venn 2-5 group data
}
\examples{
#
}
