% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.plot.R
\name{DEGp_Dotplot}
\alias{DEGp_Dotplot}
\title{Dotplot}
\usage{
DEGp_Dotplot(
  df,
  title = "xxx",
  resultdir,
  filemark,
  pic.save = T,
  pw = 8,
  ph = 10
)
}
\arguments{
\item{df}{dataframe, from DEGp_prepareDotplot result}

\item{title}{character}

\item{resultdir}{cahracter, picture save dir}

\item{filemark}{character, picture filename mark}

\item{pic.save}{logical, save picutre or not?}

\item{pw}{number, picuter width}

\item{ph}{number, picuter height}
}
\value{
#
}
\description{
Dotplot of go or kegg result(DEGp_prepareDotplot result)
}
\examples{
#
}
\author{
Jiang
}
