% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.plot.R
\name{DEGp_Dotplot2}
\alias{DEGp_Dotplot2}
\title{Dotplot2}
\usage{
DEGp_Dotplot2(
  df,
  head = 20,
  delete = NULL,
  method = 1,
  title = "xxx",
  resultdir,
  filemark,
  pic.save = T,
  pw = 8,
  ph = 10,
  pcompress = 1,
  size1 = 3,
  size2 = 5,
  color1 = "#AC001D",
  color2 = "#888888"
)
}
\arguments{
\item{df}{data.frame the dataframe must contain colname : "Description", "GeneRatio", "pvalue", "qvalue", "Count"}

\item{head}{numeric deafault is 20：keep the head 20 row of df, if not set, use all of the df}

\item{delete}{numeric or NULL delete = c(4, 7)：delete 4, 7, NULL: not delete}

\item{method}{number picture plot method, can be set one of 1:4, default is 1}

\item{title}{character the picture title}

\item{resultdir}{character outdir name where the picture save}

\item{filemark}{character the picture filename, default is "xxx"}

\item{pic.save}{logical save the picture or not,, default is T}

\item{pw}{number the picture width, default is 8}

\item{ph}{number the picture height, default is 10}

\item{pcompress}{number create a blank area at the top of the picture, default is 1}

\item{size1}{number the dot size start, default is 3}

\item{size2}{number the dot size end, default is 5}

\item{color1}{character color rgb}

\item{color2}{character color rgb}
}
\value{
ggplot2 picutre
}
\description{
Dotplot of go or kegg result(clusterProfiler enrich result) or your own data\cr
the dataframe must contain colname : "Description", "GeneRatio", "pvalue", "qvalue", "Count"\cr
note: the col GeneRatio data type should be numeric, if not please trans to numeric \cr
this function do not need prepare data  using DEGp_prepareDotplot()
}
\author{
Jiang
}
