% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.plot.R
\name{DEGp_GOALL_bar}
\alias{DEGp_GOALL_bar}
\title{GO all barplot}
\usage{
DEGp_GOALL_bar(
  GOlist.df,
  resultdir,
  filemark,
  title = "GO",
  x.which = "pvalue",
  pic.save = F,
  width = 10,
  height = 10
)
}
\arguments{
\item{GOlist.df}{dataframe, GO result}

\item{resultdir}{character, the picture outdir}

\item{filemark}{character, the picture filename mark}

\item{title}{character, the picture title}

\item{x.which}{character, x axis which to use, "pvalue" or "Count"}

\item{pic.save}{logical, save picture or not. if F, the param width and height will not work}

\item{width}{number, picture width}

\item{height}{number, picture height}
}
\value{
ggplot2 picture
}
\description{
GO all barplot
}
\author{
Jiang
}
