% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.plot.R
\name{DEGp_GOALL_barplot}
\alias{DEGp_GOALL_barplot}
\title{GO all barplot}
\usage{
DEGp_GOALL_barplot(GOlist.df, resultdir, filemark, pic.save = T)
}
\arguments{
\item{resultdir}{character, the picture outdir}

\item{filemark}{character, the picture filename mark}

\item{pic.save}{logic, save picture or not}

\item{GOlisit}{list, GO result}
}
\value{
#
}
\description{
GO all barplot
}
\examples{
#
}
