% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{DEGp_GSEA}
\alias{DEGp_GSEA}
\title{GSEA plot}
\usage{
DEGp_GSEA(GSEA.obj, num = 1)
}
\arguments{
\item{GSEA.obj}{obeject result of  clusterProfiler::GSEA()}

\item{num}{number the row postion of result of  clusterProfiler::GSEA()}
}
\value{
plot
}
\description{
GSEA plot for a result of clusterProfiler::GSEA()
}
\author{
Jiang
}
