% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{DEGp_GSEA_plotALL}
\alias{DEGp_GSEA_plotALL}
\title{GSEA plot loop}
\usage{
DEGp_GSEA_plotALL(GSEA.obj, result_dir, xl_filename)
}
\arguments{
\item{GSEA.obj}{object result of clusterProfiler::GSEA() using lots pathway gmt files}

\item{result_dir}{character the outdir name}

\item{xl_filename}{character the out filename}
}
\description{
Plot all pictures from a result of clusterProfiler::GSEA()
}
\author{
Jiang
}
