% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.plot.R
\name{DEGp_Pheat}
\alias{DEGp_Pheat}
\title{Heatmap}
\usage{
DEGp_Pheat(
  df,
  gene,
  pheno,
  outdir,
  f_mark = "",
  nlevel = 8,
  pic_w = 6,
  pic_h = 8,
  c_w = 1,
  c_h = 10,
  f_z = 12,
  f_z.col = 8,
  angle = "90",
  cluster_rows = T,
  cluster_cols = F
)
}
\arguments{
\item{df}{dataframe or matrix, the exprs data(rowname is gene name)}

\item{gene}{character, the gene name that heatmap show}

\item{pheno}{dataframe, the group info}

\item{outdir}{character, dir name}

\item{f_mark}{character, special part of filename}

\item{nlevel}{number, the group's factor level}

\item{pic_w}{number, picture width}

\item{pic_h}{number,  picture height}

\item{c_w}{number,  ceil width}

\item{c_h}{number, ceil height}

\item{f_z}{number, font size}

\item{f_z.col}{number, col font size}

\item{angle}{character, col angle}
}
\value{
#
}
\description{
plot heatmap with selected gene
}
\examples{
#
}
