% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.plot.R
\name{DEGp_prepareDotplot}
\alias{DEGp_prepareDotplot}
\title{Prepare Dotplot}
\usage{
DEGp_prepareDotplot(df, head = 20, delete = NULL)
}
\arguments{
\item{df}{dataframe, from go or kegg analysis}

\item{head}{numeric or NULL, head = 20：filter the head 20 row of df, head = NULL：all of df}

\item{delete}{numeric or NULL, delete = c(4, 7)：delete 4, 7, NULL: not delete}
}
\value{
#
}
\description{
Prepare Dotplot
}
\examples{
# 
}
\author{
Jiang
}
