% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.plot.R
\name{DEGp_prepareGOALL}
\alias{DEGp_prepareGOALL}
\title{Prepare data for DEGp_GOALL_barplot}
\usage{
DEGp_prepareGOALL(data, n_type = "ONTOLOGY", n_qu = "pvaule")
}
\arguments{
\item{data}{dataframe, GO ALL dataframe}

\item{n_type}{charater, a colname of data, the type of GO ALL dataframe, such as n_type='ONTOLOGY'}

\item{n_qu}{character, a colname of data, sort by this param, such as n_qu='pvaule'}
}
\value{
#
}
\description{
Prepare data for DEGp_GOALL_barplot. Take top n number of each type in the dataframe.
}
\examples{
#
}
\author{
Jiang
}
