% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{DEGp_prepareVolcano}
\alias{DEGp_prepareVolcano}
\title{DEGp_prepareVolcano}
\usage{
DEGp_prepareVolcano(
  resdf,
  filterc = "pvalue",
  pvalue = 0.05,
  fdr = 0.1,
  logfc.p = 1,
  logfc.n = -1,
  ymax = NULL
)
}
\arguments{
\item{resdf}{data.frame, the result of DEG analysis, must contain Gene, log2FC, PValue, FDR columns.}

\item{filterc}{character, "pvalue" or "fdr" or "both", which column to use as filter.}

\item{pvalue}{numeric, the pvalue threshold.}

\item{fdr}{numeric, the fdr threshold.}

\item{logfc.p}{numeric, the positive log2FC threshold.}

\item{logfc.n}{numeric, the negative log2FC threshold.}

\item{ymax}{numeric, the y axis threshold, the data greater ymax will be delete.}
}
\value{
data.frame, the data for the Volcano plot.
}
\description{
This function is used to prepare the data for the Volcano plot.
}
\author{
Jiang
}
