% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.plot.R
\name{DEGp_Volcano}
\alias{DEGp_Volcano}
\title{Valcano with mask gene}
\usage{
DEGp_Volcano(result, logFC = 1, adj_P = 0.2, label_geneset = NULL)
}
\arguments{
\item{result}{dataframe, all info of DEG analysis}

\item{logFC}{logfc, default 1}

\item{adj_P}{padj, default 0.2}

\item{label_geneset}{character, the mask gene list}
}
\value{
#
}
\description{
plot valcano with mask gene
}
\examples{
#
}
