% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{DEGp_Volcano_cor}
\alias{DEGp_Volcano_cor}
\title{DEGp_Volcano_cor}
\usage{
DEGp_Volcano_cor(
  resdf,
  filterc = "pvalue",
  pvalue = 0.05,
  fdr = 0.1,
  logfc.p = 1,
  logfc.n = -1,
  ymax = NULL,
  ylimup = 5,
  label_geneset = NULL,
  label_move = 0.2,
  outdir = "result",
  filename.base = "DEG_xx",
  xlab_text = "Pearson correlation(Pearson test)",
  pic.w = 7,
  pic.h = 7,
  color.not = "#BEBEBE",
  color.up = "#AC001D",
  color.down = "#3D71AD",
  size = 1.5,
  shape = 16,
  alpha = 0.9
)
}
\arguments{
\item{resdf}{data.frame, the result of DEG analysis, must contain Gene, log2FC, PValue, FDR columns.}

\item{filterc}{character, "pvalue" or "fdr", which column to use as filter.}

\item{pvalue}{numeric, the pvalue threshold.}

\item{fdr}{numeric, the fdr threshold.}

\item{logfc.p}{numeric, the positive log2FC threshold.}

\item{logfc.n}{numeric, the negative log2FC threshold.}

\item{ymax}{numeric, the y axis threshold, the data greater ymax will be delete.}

\item{ylimup}{numeric, the y axis limit up.}

\item{label_geneset}{character(n), the gene set to label.}

\item{label_move}{numeric, the move distance of the label.}

\item{outdir}{character, the output directory.}

\item{filename.base}{character, the output file name.}

\item{xlab_text}{character, the x axis label.}

\item{pic.w}{numeric, the width of the plot.}

\item{pic.h}{numeric, the height of the plot.}

\item{color.not}{character, the color of the data that is not significant.}

\item{color.up}{character, the color of the data that is up-regulated.}

\item{color.down}{character, the color of the data that is down-regulated.}

\item{size}{numeric, the size of the point.}

\item{shape}{numeric, the shape of the point.}

\item{alpha}{numeric, the transparency of the point.}
}
\value{
list, the list contains the plot and the data.
}
\description{
This function is used to plot the Volcano plot.
}
\author{
Jiang
}
