% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEGres_ToRICH}
\alias{DEGres_ToRICH}
\title{Transformat DEG_edgeR obj TO res obj}
\usage{
DEGres_ToRICH(diffan.obj, p, q, f, mark, outdir)
}
\arguments{
\item{diffan.obj}{DEG_edgeR obj}

\item{p}{pavlue}

\item{q}{fdr is q is set NULL, mean that fdr is not included for follow analysis}

\item{f}{logFC}

\item{mark}{outfile filename, if mark or outdir is not set then the result will not be written to local disk.}

\item{outdir}{outfile dirif mark or outdir is not set then the result will not be written to local disk.}
}
\value{
list(DIFF.ALL = res, DEG.ALL = genelist.deg,
             UP = genelist.up, DOWN = genelist.down,
             gsealist = genelist)
        DIFF.ALL: all deg table，genelist(deg,up,down)：genelist for enrichment
        gsealist：data for gsea analysis
}
\description{
transformat DEG_edgeR obj TO res obj for follow analysis
}
\examples{
#res <- DEG_edgeR.filter(diffobj = diffan.d, p = pvalue, q=fdr, f=log2(fc),
#                        mark = out_mark2, outdir = z_result2)
#
}
