% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSOFT.R
\name{doPerm}
\alias{doPerm}
\title{do permutations}
\usage{
doPerm(perm, X, Y, absolute, abs_method)
}
\arguments{
\item{perm}{}

\item{X}{}

\item{Y}{}

\item{absolute}{}

\item{abs_method}{}
}
\value{
a list
}
\description{
do permutations
}
