% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{find_pathway}
\alias{find_pathway}
\title{Find a pathway gmt}
\usage{
find_pathway(your_pathway = "^Ferr")
}
\arguments{
\item{your_pathway}{character the taget pathway name, support regular expression}
}
\value{
gmt
}
\description{
Find a pathway from the gmt consist of lots of pathway\cr
NOTE: this function's result is still a gmt(dataframe)
}
\author{
Jiang
}
