% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{fpkmToTpm}
\alias{fpkmToTpm}
\title{fpkm to tmp}
\usage{
fpkmToTpm(fpkm)
}
\arguments{
\item{fpkm}{numeric the data should be no log data}
}
\value{
tpm
}
\description{
fpkm to tmp
}
\examples{
# fpkmToTpm()
}
