% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode.v22.annot}
\alias{gencode.v22.annot}
\title{the genecode v22 annot and hugo annot}
\format{
A `list`.
}
\description{
the genecode v22 annot and hugo annot,
use data(gencode.v22.annot) to load it, its name: all_anot.
}
\examples{
data(gencode.v22.annot)
str(gencode.v22.annot)
}
