% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_search_expr.R
\name{Gene_dens}
\alias{Gene_dens}
\title{Plot Gene Expression Density}
\usage{
Gene_dens(df, tagetGene)
}
\arguments{
\item{df}{data.frame, the data.frame of the clinical and gene expression data}

\item{tagetGene}{character, the gene name in the data.frame}
}
\value{
ggplot2 object
}
\description{
Plot Gene Expression Density
}
\examples{
#
}
