% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{getData}
\alias{getData}
\title{Get Data}
\usage{
getData(..., package = "LZ")
}
\arguments{
\item{...}{character, the alias of bulit-in dataset, ex. gencode.v22.annot
keggpathway.gmt}

\item{package}{character, package name}
}
\value{
list Large list
}
\description{
Get Data
}
\examples{
#  data(gencode.v22.annot, package = "LZ")
#  data(keggpathway.gmt, package = "LZ")
}
\author{
Jiang
}
