% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGapi.R
\name{getKEGGcpd.loop}
\alias{getKEGGcpd.loop}
\title{Get KEGG compound Loop}
\usage{
getKEGGcpd.loop(pathway)
}
\arguments{
\item{pathway}{character(n), name character the name of `pathway` is the hsa_id}
}
\value{
dataframe
}
\description{
获取所有人类hsa通路中的代谢物(通过循环实现)
}
\author{
Jiang
}
