% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGapi.R
\name{getKEGGcpd}
\alias{getKEGGcpd}
\title{Get KEGG compound}
\usage{
getKEGGcpd(id)
}
\arguments{
\item{id}{character(1), hsa_id such as "hsa00010", ...}
}
\value{
dataframe
}
\description{
获取指定人类hsa通路中的代谢物（通过R KEGGREST实现）
}
\author{
Jiang
}
