% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gmt_longTowide2}
\alias{gmt_longTowide2}
\title{GMT Long to Wide data use groub_by}
\usage{
gmt_longTowide2(gmtdf)
}
\arguments{
\item{gmtdf}{dataframe the long dataframe consist of two coloumns.\cr
# term        gene
# pathway1   gene1
# pahtway1   gene2
# ...        ...
# pahtwayn   gene1
# pathwayx   genex}
}
\value{
gmtdf.w
}
\description{
transformat long GMT dataframe to wide dataframe
}
\examples{
# return datafame as this:
# term       gene
# pathway1   gene1,gene2,...
# pathway2   gene1,genex,...
}
\author{
Jiang
}
