% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gmt_longTOwide3}
\alias{gmt_longTOwide3}
\title{GMT Long to Wide data}
\usage{
gmt_longTOwide3(df)
}
\arguments{
\item{df}{dataframe the long dataframe consist of two coloumns.\cr
# term        gene
# pathway1   gene1
# pahtway1   gene2
# ...        ...
# pahtwayn   gene1
# pathwayx   genex}
}
\value{
gmt.wide
}
\description{
trans long GMT dataframe to wide dataframe
}
\examples{
# return datafame as this:
# term       gene
# pathway1   gene1,gene2,...
# pathway2   gene1,genex,...
}
\author{
Jiang
}
