% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{GSEA_plotALL}
\alias{GSEA_plotALL}
\title{GSEA plot loop}
\usage{
GSEA_plotALL(gsea.all, result_dir, filename)
}
\arguments{
\item{gsea.all}{object result of clusterProfiler::GSEA() using lots pathway gmt files}

\item{result_dir}{character the outdir name}

\item{filename}{character the out filename}
}
\description{
Plot all pictures from a result of clusterProfiler::GSEA()
}
\author{
Jiang
}
