% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune.R
\name{immuneScore_clean}
\alias{immuneScore_clean}
\title{immuneScore_clean}
\usage{
immuneScore_clean(immuneScore_df, method)
}
\arguments{
\item{immuneScore_df}{data frame, the result of immuneScore()}

\item{method}{character, one of "mcp_counter", "quantiseq", "epic", "xcell", "estimate", "timer", "cibersort", "ssGSEA"}
}
\value{
a data frame
}
\description{
Clean the immuneScore() result
}
\examples{
\dontrun{
res <- immuneScore(res_mcp_counter, method = "mcp_counter")
}
}
