% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune.R
\name{immuneScore_onestep}
\alias{immuneScore_onestep}
\title{immuneScore_onestep}
\usage{
immuneScore_onestep(exprs, tcga_abbr = "brca", value = T, QN = F, rapid = F)
}
\arguments{
\item{exprs}{matrix, gene expression matrix, NOT log-transformed}

\item{tcga_abbr}{character, TCGA abbreviation, only used for method = "timer"}

\item{value}{logical, default T, if T, assign variables to global environment}

\item{QN}{logical, default F, RNAseq data: recommend QN = F, otherwise T, only used for method = "cibersort"}

\item{rapid}{logical, default F, if T, only run mcp_counter, quantiseq, epic, estimate, ssGSEA, xcell, timer\cr
otherwise, also run cibersort and cibersort_abs(will take a long time)}
}
\value{
a data frame
}
\description{
Calculate immune score using RNAseq data or mcroarray data using multiple methods in one step
}
\examples{
\dontrun{
res_df <- immuneScore_onestep(exprs, tcga_abbr = "brca")
}
}
