% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune.R
\name{immuneScore_test}
\alias{immuneScore_test}
\title{immuneScore_test}
\usage{
immuneScore_test(immuneScore_df, group_df, test_method = "wilcox")
}
\arguments{
\item{immuneScore_df}{data.frame, the result of immuneScore() or immuneScore_onestep()}

\item{group_df}{data.frame, the group information, must be 2-coloum-data.frame, the first column is the ID, the second column is the group}

\item{test_method}{character, default "wilcox", one of "wilcox", "t-test".}
}
\value{
a list
}
\description{
Test the difference of immune score between two groups
}
\examples{
#
}
