% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.R
\name{install}
\alias{install}
\title{install package by using CRAN or bioconductor}
\usage{
install(package)
}
\arguments{
\item{package}{charater the package name you want to install}
}
\value{
#
}
\description{
install package by using CRAN or bioconductor, if the package is not in
the CRAN, then the function will use bioconductor method to install the package.
}
\examples{
#
# packs <- c('dplyr', 'stringr', 'DESeq2')
# for (p in packs) {
#   LZ::install(p)
# }
# or
# LZ::install("ggrepel")
}
\author{
Jiang
}
