% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{lz.annot}
\alias{lz.annot}
\title{Annotate expression data}
\usage{
lz.annot(edata, annotdf, row.names = T)
}
\arguments{
\item{edata}{matrix or dataframe, expression data, suggest contrain only numeric columns \cr
or dataframe that the first column is character vector(the column name is 'ID')}

\item{annotdf}{dataframe, annotation data, suggest contrain only two columns: 'ID' and 'Symbol'}

\item{row.names}{logical, default is TRUE, if T, all the column must be numeric, \cr
if F, the first column must be character vector.}
}
\value{
dataframe, annotated expression data
}
\description{
Annotate expression data
}
\author{
Jiang
}
