% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{lz_isna}
\alias{lz_isna}
\title{find the NA(NA, grepl("\\s|^$", lie))}
\usage{
lz_isna(lie)
}
\arguments{
\item{lie}{vector the column or vector of your data}
}
\value{
vector
}
\description{
find the NA(NA, grepl("\\s|^$", lie))
}
\examples{
# lz_isna()
}
