% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{lzhead}
\alias{lzhead}
\title{show part info of a dataframe}
\usage{
lzhead(df, ncol = 4, nrow = 3, rowstart = 1, colstart = 1)
}
\arguments{
\item{df}{dataframe data}

\item{ncol}{number default is 4, the end colomn index}

\item{nrow}{number default is 3, the end colomn index}

\item{rowstart}{number default is 1, the start row index}

\item{colstart}{number default is 1, the start colomn index}
}
\value{
#
}
\description{
show part info of a dataframe
}
\examples{
# df \%>\% lzhead
}
