% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{min.max.scaler}
\alias{min.max.scaler}
\title{Normalize the vector to 0-1}
\usage{
\method{min}{max.scaler}(vec)
}
\arguments{
\item{vec}{numeric(n), the vector need to be normalized}
}
\value{
numeric(n) the normalized vector
}
\description{
Normalize the vector to 0-1
}
\author{
Jiang
}
