% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{mkdir.p}
\alias{mkdir.p}
\title{Create a folder, regardless of whether the folder already exists}
\usage{
mkdir.p(dir)
}
\arguments{
\item{dir}{character, the dir name}
}
\value{
# Create a folder regardless of whether the folder already exists.
}
\description{
Create a folder, regardless of whether the folder already exists, overwrite if it exists, in deg analysis usually used to make output dir.
}
\examples{
#
}
