% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{mkdir}
\alias{mkdir}
\title{create dir}
\usage{
mkdir(dir)
}
\arguments{
\item{dir}{character, the dir name}
}
\value{
# if the dir no exist the function will create the dir
}
\description{
mkdir, in deg analysis usually used to make output dir
}
\author{
Jiang
}
