% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_prepareData.S3.R
\name{newDEeset}
\alias{newDEeset}
\title{Create a new DEeset object}
\usage{
newDEeset(eset, group, f_mark = "DE")
}
\arguments{
\item{eset}{dataframe, the cleaned exprs dataframe, rowname is gene_id}

\item{group}{dataframe, the cleaned group dataframe, rowname is sample, \cr
the first cloumn is the Type and must be factor}

\item{f_mark}{character, the mark of this DEeset object, default is 'DE'}
}
\value{
a DEeset object
}
\description{
Create a new DEeset object by exsited eset and group object
}
\author{
Jiang
}
