% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_prepareData.S3.R
\name{newXENA}
\alias{newXENA}
\title{Create a new XENA object}
\usage{
newXENA(eset.count, info)
}
\arguments{
\item{eset.count}{dataframe, the exprs dataframe, column is gene_id\cr
the last column is the sample, row is the gene.}

\item{info}{character, the information of eset.count}
}
\value{
a XENA object
}
\description{
Create a new XENA object
}
