% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.R
\name{p_install}
\alias{p_install}
\title{install package by using CRAN or bioconductor}
\usage{
p_install(packs)
}
\arguments{
\item{packs}{charater(n), the package name you want to install}
}
\value{
no
}
\description{
install package by using CRAN or bioconductor, if the package is not in
the CRAN, then the function will use bioconductor method to install the package.
}
\author{
Jiang
}
