% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.R
\name{require_pack}
\alias{require_pack}
\title{check if the package is been installed}
\usage{
require_pack(packname)
}
\arguments{
\item{packname}{character the R package name}
}
\value{
# message in console
}
\description{
check if the package is been installed
}
\author{
Jiang
}
