% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_gsea.R
\name{runAPP_GSEA}
\alias{runAPP_GSEA}
\title{Run Shiny APP GSEA}
\usage{
runAPP_GSEA(genelist)
}
\arguments{
\item{genelist}{list, a name list, gene list sorted by log2FC}
}
\value{
no
}
\description{
Run Shiny APP GSEA
}
\author{
Jiang
}
