% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{scale_minmax}
\alias{scale_minmax}
\title{Normalize the vector to 0-1}
\usage{
scale_minmax(vec)
}
\arguments{
\item{vec}{numeric(n), the vector need to be normalized}
}
\value{
normalized vector
}
\description{
Normalize the vector to 0-1
}
\author{
Jiang
}
