% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_km.R
\name{Surv_diffONE}
\alias{Surv_diffONE}
\title{Survival analysis and plot}
\usage{
Surv_diffONE(
  eset_os,
  tagetGene,
  method = "median",
  num.tran = 365,
  break.x = 2.5,
  col_name = "Group",
  color = c("#eaa8a8", "#729fc9")
)
}
\arguments{
\item{eset_os}{data.frame, the data.frame of the clinical and gene expression data}

\item{tagetGene}{character, the gene name}

\item{method}{character, the method to cut the gene expression data, \cr
can be "mean", "upper75 vs low25", "median", "upper25 vs low75", "auto cut point", "upper25 vs low25" or a number}

\item{num.tran}{number, the number to transform the OS.time, mean: OS.time/num.tran}

\item{break.x}{number, the break x}

\item{col_name}{character, the column name of the group column, default is "Group"}

\item{color}{character, the color of the plot}
}
\value{
a list
}
\description{
Survival analysis and plot using multi methods
}
\examples{
\dontrun{
eset_os
     OS.time OS  EGFR  TP53 ...
P1     3450  0    4.6   6.6 ...
P2     1280  1    1.0   6.0 ...
P3      455  1    3.0   4.0 ...
res <- Surv_diffONE(eset_os, tagetGene="EGFR", method = "median",
                    num.tran=365, break.x=2.5, col_name="Group",
                    color=c("#eaa8a8", "#729fc9"))
}
}
