% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_km.R
\name{Surv_ostime_cut}
\alias{Surv_ostime_cut}
\title{ostime cut using different methods}
\usage{
Surv_ostime_cut(eset_os.df, tagetGene, method = "median")
}
\arguments{
\item{eset_os.df}{data.frame, the data.frame of the clinical and gene expression data}

\item{tagetGene}{character, the gene name, the column name of the gene expression data}

\item{method}{character, default "mean", the method to cut the gene expression data, \cr
can be "mean", "upper75 vs low25", "median", "upper25 vs low75", "auto cut point", "upper25 vs low25" or a number}
}
\value{
a list
}
\description{
Use different methods to cut the data into two groups and then fuse this grouping information into data
}
\examples{
\dontrun{
eset_os.df
     OS.time OS  EGFR  TP53 ...
P1     3450  0    4.6   6.6 ...
P2     1280  1    1.0   6.0 ...
P3      455  1    3.0   4.0 ...
}
}
