% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_gsea.R
\name{syGSEA}
\alias{syGSEA}
\title{Shiny APP GSEA}
\usage{
syGSEA(gmt.largelist, genelist)
}
\arguments{
\item{gmt.largelist}{dataframe or gmt file from msigdb}

\item{genelist}{list, a name list, gene list sorted by log2FC}
}
\value{
no
}
\description{
Shiny APP GSEA
}
\author{
Jiang
}
