% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{gencode.v22.annot}
\alias{KEGGPATHWAYdb}
\alias{gmt.largelist.23.12.Hs.symbols}
\title{Datasets
gencode.v22.annot contains gene annot info.}
\description{
Datasets
gencode.v22.annot contains gene annot info.

Datasets
KEGGPATHWAYdb contains kegg pathway information

Datasets
gmt.largelist.23.12.Hs.symbols contains human msigdb pathway and kegg pathway symbols information
}
\keyword{datasets}
