% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{gencode.v22.annot}
\alias{KEGGPATHWAYdb}
\alias{gmt.largelist.23.12.Hs.symbols}
\alias{annot_hugo}
\alias{LM22}
\title{Datasets
gencode.v22.annot contains gene annot info.}
\description{
Datasets
gencode.v22.annot contains gene annot info.

Datasets
KEGGPATHWAYdb contains kegg pathway information

Datasets
gmt.largelist.23.12.Hs.symbols contains human msigdb pathway and kegg pathway symbols information

Datasets
annot_hugo contains human gene annot and type information

Datasets
LM22 contains 22 immune cell types gene signature
}
\keyword{datasets}
