% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEG_DESeq2.dds}
\alias{DEG_DESeq2.dds}
\title{RNAseq differential analysis DESeq2 001<dds>}
\usage{
DEG_DESeq2.dds(exprset.group, batch = F)
}
\arguments{
\item{exprset.group}{the list consisted of eset, group and mark.}

\item{batch}{batch number}
}
\value{
dds ( DESeq(dds, parallel = T) )
}
\description{
The function is designed to perform differential analysis by DESeq2<dds>.
}
\examples{
# exprset.group = list(eset, group, f_mark)
# ** 如果bacth=T，则group中第二列为batch列。
# ** 矩阵数据格式(数值型，整型)
#       | row1 | row2 | row3  | row4
# gene1 |  34  |  23  |  56   |  23
# gene2 |  35  |  23  |  12   |  23
# gene3 |  12  |  78  |  78   |  78
# ** 分组数据格式：需要组的行名=表达谱的列名 rownames(group) == colname(eset)
#            Type   BATCH
# rowname1 | tumor  1
# rowname2 | tumor  1
# rowname3 | normal 2
# rowname4 | normal 2
}
