% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.R
\name{DEG_runENRICH}
\alias{DEG_runENRICH}
\title{run KEGG analysis inFCs(2, 1.5) in batches(upgene, downgene, allgene)}
\usage{
DEG_runENRICH(
  genelist,
  outdir,
  glist.save = T,
  rungo = T,
  runkegg = T,
  rapid = T
)
}
\arguments{
\item{genelist}{list, must be two-dimensional name list, result of sapply DEG_prepareGOglist}

\item{outdir}{character, the output dir}

\item{glist.save}{logical default is TRUE, wether save genelist}

\item{rungo}{logical default is TRUE, wether run go analysis}

\item{runkegg}{logical default is TRUE, wether run kegg analysis}

\item{rapid}{loggicla defaut is F, if set F, the function will plot go tree picture, which make cost much time.}
}
\value{
# GO object & write result to xlsx
}
\description{
run GO analysis in batches(upgene, downgene, allgene) and write result to xlsx
}
\examples{
#
}
\author{
Jiang
}
