% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEG_voom}
\alias{DEG_voom}
\title{RNAseq differential analysis Limma voom}
\usage{
DEG_voom(exprset.group, pval = 0.05, fdr = 0.1, logfc = 1)
}
\arguments{
\item{exprset.group}{the list consisted of eset, group and mark.}

\item{pval}{default 0.05}

\item{fdr}{default 0.1}

\item{logfc}{default 1}
}
\value{
list(groupdata = pheno, resdf=tab, deg=deg)
}
\description{
The function is designed to perform differential analysis by Limma voom.
}
\examples{
#
}
