% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.enrich.plot.R
\name{DEGp_GOALL_bar}
\alias{DEGp_GOALL_bar}
\title{GO all barplot}
\usage{
DEGp_GOALL_bar(GOlist.df, resultdir, filemark, pic.save = T)
}
\arguments{
\item{GOlist.df}{dataframe, GO result}

\item{resultdir}{character, the picture outdir}

\item{filemark}{character, the picture filename mark}

\item{pic.save}{logic, save picture or not}
}
\value{
#
}
\description{
GO all barplot
}
\author{
Jiang
}
