% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.plot.R
\name{DEGp_Pheat}
\alias{DEGp_Pheat}
\title{Heatmap}
\usage{
DEGp_Pheat(
  df,
  gene,
  pheno,
  outdir,
  f_mark = "",
  nlevel = 8,
  pic_w = 6,
  pic_h = 8,
  c_w = 6,
  c_h = 10,
  c_colors = c("blue", "#FFFFCC", "red"),
  color.num = 100,
  fontsize = 12,
  fontsize.col = 8,
  angle = "90",
  cluster_rows = T,
  cluster_cols = F,
  clustering_method = "complete",
  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean"
)
}
\arguments{
\item{df}{dataframe or matrix, the exprs data(rowname is gene name)}

\item{gene}{character, the gene name that heatmap show}

\item{pheno}{dataframe, the group info}

\item{outdir}{character, dir name}

\item{f_mark}{character, special part of filename}

\item{nlevel}{number, the group's factor level}

\item{pic_w}{number, picture width}

\item{pic_h}{number,  picture height}

\item{c_w}{number,  ceil width}

\item{c_h}{number, ceil height}

\item{c_colors}{character(n), default isc ("blue", "#FFFFCC", "red")}

\item{color.num}{number, default is 100}

\item{fontsize}{number, row(sample) annot font size}

\item{fontsize.col}{number, col(gene) annot font size}

\item{angle}{character, col angle}

\item{cluster_rows}{logical, default is T}

\item{cluster_cols}{logical, default is F}

\item{clustering_method}{character, one of c("ward.D", "ward.D2", "single", "complete", 
"average", "mcquitty", "median", "centroid")}

\item{clustering_distance_rows}{character, one of c("correlation", "euclidean", "maximum", 
"manhattan", "canberra", "binary", "minkowski")}

\item{clustering_distance_cols}{character, as the same to `clustering_distance_rows`}
}
\value{
#
}
\description{
plot heatmap with selected gene
}
\examples{
\dontrun{
  groupdf <- data.frame(row.names = colnames(sig.expr),
                        Type = c(rep("11A", 3), rep("01A", 3)),
                        check.names = F)
  groupdf[,1] <- as.factor(groupdf[,1])
  m <- c("ward.D", "ward.D2", "single", "complete", 
          "average", "mcquitty", "median", "centroid")
  d <- c("correlation", "euclidean", "maximum", "manhattan",
         "canberra", "binary", "minkowski")
  for (i in m) {
    for (i.d in d) {
      cat(i, i.d, "\n")
      DEGp_Pheat(
        sig.expr,
        gene = show.gene, 
        pheno = groupdf,
        outdir = "D:/xxx/yyy",
        f_mark = paste0("heatmap.", i, "_", i.d),
        nlevel = 2,
        pic_w = 5,
        pic_h = 8,
        c_w = 18,
        c_h = 10,
        fontsize = 10,
        fontsize.col = 14,
        angle = "90",
        cluster_rows = T,
        cluster_cols = T,
        clustering_method = i,
        clustering_distance_rows=i.d
      )  
    }
  }
}
}
