% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.plot.R
\name{DEGp_prepareVolcano}
\alias{DEGp_prepareVolcano}
\title{prepare the data that valcano need}
\usage{
DEGp_prepareVolcano(df_valcano, filterc = "fppadj")
}
\arguments{
\item{df_valcano}{dataframe, the all data of DEG analysis}

\item{filterc}{character, the way of show picture}
}
\value{
#
}
\description{
prepare the data that valcano need
}
\examples{
#
}
