% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.plot.R
\name{DEGp_Volcano2}
\alias{DEGp_Volcano2}
\title{Valcano2 with mask gene}
\usage{
DEGp_Volcano2(
  resdf,
  filterc = "both",
  pvalue = 0.05,
  fdr = 0.1,
  logfc.p = 1,
  logfc.n = -1,
  label_geneset = NULL,
  outdir = "result",
  filename.base = "DEG_xx",
  pic.w = 7,
  pic.h = 7,
  color.not = "#BEBEBE",
  color.up = "#AC001D",
  color.down = "#3D71AD",
  ymax = NULL,
  size = 1.5,
  shape = 16,
  alpha = 0.9
)
}
\arguments{
\item{resdf}{dataframe the all data of DEG analysis}

\item{filterc}{character the way of show picture, can be one of "both", "pvalue", "fdr"}

\item{pvalue}{number pvalue, default 0.05}

\item{fdr}{number fdr, default 0.1}

\item{logfc.p}{number log2fc positive, default 1}

\item{logfc.n}{number log2fc negative, default -1}

\item{label_geneset}{character(n) the genelist need be shown in the valcano, default NULL}

\item{outdir}{character the picture outdir, default "result"}

\item{filename.base}{character the picture filename base, default "DEG_xx"}

\item{pic.w}{number the picture width, default 7}

\item{pic.h}{number the picture height, default 7}

\item{color.not}{character the picture color to not sig gene, default "#BEBEBE"}

\item{color.up}{character the picture color to up sig gene, default "#AC001D"}

\item{color.down}{character the picture color to down sig gene, default "#3D71AD"}

\item{ymax}{number the max ylim, default NULL}

\item{size}{number the picture point size, default 1.5 ...}

\item{shape}{number the picture point shape, default 16, or 1 ...}

\item{alpha}{number the transpate of picture, default 0.9}
}
\value{
list
}
\description{
plot valcano with mask gene
}
\author{
Jiang
}
