% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{DEGres_ToGSEA}
\alias{DEGres_ToGSEA}
\title{Transformat DEG_edgeR obj to GSEA APP files}
\usage{
DEGres_ToGSEA(diffan.obj, outdir)
}
\arguments{
\item{diffan.obj}{DEG_edgeR obj}

\item{outdir}{the result output dir}
}
\value{
write to local disk without anything retrun to object
}
\description{
transformat DEG_edgeR obj to the files that GSEA APP need
}
\examples{
# suppressWarnings({
# DEG_resTogesa(diffan.obj = diffan.d,
#               outdir = z_result4)
# })
}
