% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{limma.general}
\alias{limma.general}
\title{General compare the differnt ID(Gene) between the group B vs A}
\usage{
limma.general(eset, group, pval = 0.05, fdr = 0.1, logfc = log2(2))
}
\arguments{
\item{eset}{data.frame, the exprs data, must be log2 transformed}

\item{group}{data.frame, the pheno information, the first coloumn is the type information and must be factor.}

\item{pval}{number, the pvalue cutoff}

\item{fdr}{number, the fdr cutoff}

\item{logfc}{number, the log2(fc) value cutoff}
}
\value{
# all deg dataframe(resdf)
}
\description{
General compare the differnt ID(Gene) between the group B vs A
}
\examples{
\dontrun{
pn_eset <- data.table::fread("./protein_count_pn.csv", data.table = F)
pn_eset <- pn_eset \%>\% tibble::column_to_rownames(var = "Protein")

group <- read.csv("./group.csv", row.names = 1)
group$Type <- as.factor(group$Type)

df <- log2(pn_eset + 1)
diff <- limma.general(eset = df, group = group)
}
}
