% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{quchong}
\alias{quchong}
\title{delete the duplicate name of the first column of dataframe}
\usage{
quchong(
  eset,
  col.by,
  col.del = NULL,
  auto.del.character = T,
  summarise = F,
  summarise.fun = "mean",
  keep.one.fun = "mean"
)
}
\arguments{
\item{eset}{dataframe, the first column must be gene, type: character}

\item{col.by}{character, delete the duplicate by this column}

\item{col.del}{character, manually delete some column}

\item{auto.del.character}{logical, if auto delete the column those class is character}

\item{summarise}{logical, if summarise the data by col.by, if the character column contrain \cr
important information, meaning that you can't summarise some rows to one row, \cr
you should set this param to FALSE and use keep.one.fun to keep one row with not summarise.}

\item{summarise.fun}{character, the summarise function, default is mean, support mean, median, max}

\item{keep.one.fun}{character, the keep one function, default is mean, support mean, median}
}
\value{
dataframe
}
\description{
delete the duplicate name of the first column of dataframe
}
\examples{
\dontrun{
quchong(eset, col.by="id")
head(eset)[,1:3]
          gene_id cont-1 cont-2
  ENSG00000186827      0      1
  ENSG00000186891     63     50
  ENSG00000160072   1218   1023
}
}
\author{
Jiang
}
